Changed Profile format to single by reprensetation. 20 * size(char) * seq. length

+++ src/prefiltering/QueryScoreGlobal.h
@@ -13,7 +13,7 @@
 class QueryScoreGlobal : public QueryScore {
     
 public:
-    QueryScoreGlobal(int dbSize, unsigned short * seqLens, int k, short kmerThr, double kmerMatchProb, float zscoreThr);
+    QueryScoreGlobal(int dbSize, unsigned int *seqLens, int k, short kmerThr, double kmerMatchProb, float zscoreThr);
 
     ~QueryScoreGlobal();
